requirements.txt
opencv-python>=4.0.0.21
pytesseract >=0.2.6
Django>=2.1.5
django-cleanup>=3.0.1
djangorestframework>=3.9.1
Markdown>=3.0.1
